hancan

+++ Assets/Script/SkillData.cs
@@ -37,7 +37,7 @@ public class SkillData
     public DoubleList_Vector2[] specialAoe; // 특수 범위 (2차원 좌표 리스트)
     public int cooldown;        // 쿨타임
     public int colldownSkill;      // 지속 스킬횟수
-    public float basicValue;      // 기본 위력
+    public int basicValue;      // 기본 위력
     public UDictionary<IncreaseType, float> increase; // 위력 계수 (예: {공격력: 0.1}, {방어력: 0.5})
     public int damageHit;         // 타격 횟수
     public float hitSpeed;        // 타격 속도
